Fix: Button 분기처리
