+++ src/pages/_app.tsx
@@ -4,8 +4,9 @@ import type { AppProps } from "next/app";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <Header>
+    <>
+      <Header />
       <Component {...pageProps} />
-    </Header>
+    </>
   );
 }
